+++ examples/run_grpo.py
@@ -15,17 +15,14 @@
 import argparse
 import os
 import pprint
-from copy import deepcopy
-from dataclasses import dataclass
 
-import jsonlines
 from omegaconf import OmegaConf
 from transformers import AutoTokenizer
 
 from nemo_rl.algorithms.grpo import MasterConfig, grpo_train, setup
 from nemo_rl.algorithms.utils import get_tokenizer
 from nemo_rl.data import DataConfig
-from nemo_rl.data.interfaces import DatumSpec
+from nemo_rl.data.datasets import JsonlinesDataset
 from nemo_rl.distributed.virtual_cluster import init_ray
 from nemo_rl.environments.math_environment import MathEnvironment
 from nemo_rl.models.generation.interfaces import configure_generation_config
@@ -46,81 +43,6 @@ def parse_args():
     args, overrides = parser.parse_known_args()
 
     return args, overrides
-
-
-@dataclass
-class JsonlinesDataset:
-    jsonl_path: str
-    seed: int
-    tokenizer: AutoTokenizer
-    max_seq_length: int
-    filter_long_samples: bool = False
-
-    def __post_init__(self):
-        self.data = self._load_data()
-
-        idx_to_ignore = set()
-        if self.filter_long_samples:
-            for i, item in enumerate(self):
-                if item["length"] > self.max_seq_length:
-                    idx_to_ignore.add(i)
-            print(f"found {len(idx_to_ignore)} long samples to ignore on dataset init")
-
-        self.data = [item for i, item in enumerate(self.data) if i not in idx_to_ignore]
-
-    def _load_data(self):
-        with jsonlines.open(self.jsonl_path, "r") as reader:
-            data = [line for line in reader]
-        return data
-
-    def __len__(self):
-        return len(self.data)
-
-    def __getitem__(self, idx: int) -> DatumSpec:
-        data = self.data[idx]
-        # support single turn for now
-        assert len(data["messages"]) == 1
-        single_message = data["messages"][0]
-
-        message_log = []
-
-        # this will also contain system prompt
-        user_message = {"role": "user"}
-
-        for m in single_message:
-            if m["role"] == "user":
-                # need to be deepcopy to avoid overwriting the original metadata
-                extra_env_info = deepcopy(m["metadata"])
-
-        message = self.tokenizer.apply_chat_template(
-            single_message,
-            tokenize=False,
-            add_generation_prompt=True,
-            add_special_tokens=False,
-        )
-        user_message["token_ids"] = self.tokenizer.apply_chat_template(
-            single_message,
-            tokenize=True,
-            add_generation_prompt=True,
-            add_special_tokens=False,
-            return_tensors="pt",
-        )[0]
-        user_message["content"] = message
-        message_log.append(user_message)
-
-        length = sum(len(m["token_ids"]) for m in message_log)
-
-        output = {
-            "message_log": message_log,
-            "length": length,
-            "extra_env_info": extra_env_info,
-            "loss_multiplier": 1.0,
-            "idx": idx,
-            "task_name": data["task_name"],
-            "dataset": data["dataset"],
-        }
-
-        return output
 
 
 def setup_data(tokenizer: AutoTokenizer, data_config: DataConfig, env_configs):
